fix(m6-pr4b): the two 68000 MOVE emit-arm IL defects — d16(An) IL + wide-store page OOB
PR-4a made the 68000 MOVE emit arm LIVE, which surfaced two emit-arm defects the
MOVE-family parity sweep was Skip'd on (honest do-not-merge). Both are now fixed and
the sweep is re-armed as a load-bearing, byte-identical merge gate across EVERY EA
mode (incl. mode-5 d16(An) as src+dest and the A7 edges).

Bug 1 — d16(An) emitted invalid CIL (InvalidProgramException). EmitAddDisp16 typed its
sign-extended displacement as `short` and passed it to `il.Emit(OpCodes.Ldc_I4, disp)`.
ILGenerator has an `Emit(OpCode, Int16)` overload, so the `short` silently bound to it
and emitted a TRUNCATED 2-byte operand for the 4-byte-inline Ldc_I4 — a malformed IL
stream the CLR rejects at execute. Fixed by holding `disp` as an `int` (sign-extended),
selecting `Emit(OpCode, int)` (mirrors EmitAbsW). MOVEQ itself was always valid — MOVE.q
only failed because a discovered block included a d16 MOVE neighbour.

Bug 2 — wide/byte store page derivation overran the fastmem/DirtyMap page arrays
(IndexOutOfRangeException; mis-attributed to "the A7 register map" — A7 banking was
always correct). The page index `ea >> 8` did not clamp the resolved EA to the bus
AddressMask, so a MOVE whose full-32-bit A-register-relative dest carried bits above the
24-bit bus indexed past `bool[PageCount]`/the fastmem arrays. Fixed by masking the EA
with `_bus.AddressMask` before every page derivation (EmitMaskAddrLocalToBus /
EmitMaskEaLocalToBus, in EmitMarkWidePagesDirty / EmitStoreByte / LoadByteFromBus).
Identity for the byte CPUs (their EA is already in range), so 6502/Z80/8086 are
unperturbed.

Pre-merge review (HIGH): the cross-page END term `(addr + byteSpan - 1) >> 8` re-escaped
the space when the base sits at the top of the 24-bit bus (e.g. MOVE.l to 0x00FFFFFF).
Masking only the base was insufficient. EmitWideEndPage now re-clamps the end sum with
AddressMask so the trailing bytes wrap to page 0 exactly as the interpreter's per-Write8
masking does. Pinned by the deterministic Wide_MOVE_to_top_of_address_space regression
test (throws IndexOutOfRange without the mask; byte-identical with it).

Gates: build clean (0 warnings); MOVE-family sweep un-Skip'd + byte-identical across all
EA modes (sample 3000); dead-arm-now-live holds (M68kMoveEmitSelections > 0); byte-CPU
JIT sweeps green (6502/Z80/8086); FallbackEmitCount drop holds.

+++ src/CpuEmulator.Jit/BlockCompiler.M68000.cs
@@ -312,11 +312,17 @@ internal sealed partial class BlockCompiler<TCpu> where TCpu : class
     }
 
     /// <summary>d16(An): ea = An + (int)(short)ext[0]. An is on the stack (uint); add the compile-time signed
-    /// 16-bit displacement (baked Ldc_I4). Leaves ea (uint, via the unchecked Add) on the stack.</summary>
+    /// 16-bit displacement (baked Ldc_I4). Leaves ea (uint, via the unchecked Add) on the stack.
+    /// <para>PR-4b: <paramref name="disp"/> is held as an <c>int</c> (the sign-extended short), NOT a <c>short</c>.
+    /// ILGenerator has both an <c>Emit(OpCode, short)</c> (Int16) and an <c>Emit(OpCode, int)</c> overload; passing
+    /// a <c>short</c> silently binds to the Int16 overload, which writes the <c>Ldc_I4</c> opcode (0x20, a 4-byte
+    /// inline operand) but emits only a TRUNCATED 2-byte operand — a malformed IL stream the CLR rejects at execute
+    /// with <c>InvalidProgramException</c>. Sign-extending to <c>int</c> first selects <c>Emit(OpCode, int)</c> and
+    /// emits the correct operand (this mirrors <see cref="EmitAbsW"/>, which already uses an <c>int</c>).</para></summary>
     private void EmitAddDisp16(EmitContext ctx, ushort pc, ref int extIndex)
     {
-        short disp = unchecked((short)NextExtWord(pc, ref extIndex));
-        ctx.Il.Emit(OpCodes.Ldc_I4, disp);           // sign-extended to int
+        int disp = unchecked((short)NextExtWord(pc, ref extIndex));   // sign-extend short -> int (selects Emit(OpCode,int))
+        ctx.Il.Emit(OpCodes.Ldc_I4, disp);
         ctx.Il.Emit(OpCodes.Add);
     }
